line shape works
